Use `?mode` and `?eq` to pass through optional arguments.
This way the default argument for `eq` does not need to be repeated.

This also makes the `mode` argument optional.  I think `Seq` is a safe default.

Finally, when you pass an optional argument using `~` rather than `?` it is
possibly slightly slower, because it potentially involves allocating an option.
It should not make a big difference, but it is something to keep in mind at
least for very tight inner loops.